requirements.txt
ccxt==4.0.96
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
ta==0.10.2
requests==2.31.0
ta-lib==0.4.28
scipy==1.11.4
Flask==3.0.0
Flask-CORS==4.0.0
matplotlib==3.8.2
